requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
Pillow
django-taggit